+++ integration/nonmem/bbi_sge_test.go
@@ -1,15 +1,11 @@
 package nonmem
 
 import (
-	"context"
 	"fmt"
-	"os"
-	"path/filepath"
 	"testing"
 	"time"
 
 	"github.com/metrumresearchgroup/gogridengine"
-	"github.com/metrumresearchgroup/wrapt"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -18,56 +14,13 @@ func TestBbiCompletesSGEExecution(tt *testing.T) {
 		tt.Skip("Skipping SGE as it's not enabled")
 	}
 
-	tests := []struct {
-		name string
-	}{
-		{name: "acop"},
-		{name: "ctl_test"},
-		{name: "leading-path-with space"},
+	tests := []string{
+		"acop",
+		"ctl_test",
+		"leading-path-with space",
 	}
 
-	for _, test := range tests {
-		tt.Run("SGE "+test.name, func(tt *testing.T) {
-			t := wrapt.WrapT(tt)
-
-			scenario := InitializeScenario(t, test.name)
-
-			// Test shouldn't take longer than 5 min in total
-			// TODO use the context downstream in a runModel function
-			ctx, cancel := context.WithTimeout(context.Background(), 30*time.Minute)
-			defer cancel()
-
-			scenario.Prepare(t, ctx)
-
-			for _, model := range scenario.models {
-				t.Run(model.identifier, func(t *wrapt.T) {
-					nonMemArguments := []string{
-						"-d",
-						"nonmem",
-						"run",
-						"sge",
-						"--nm_version",
-						os.Getenv("NMVERSION"),
-					}
-
-					_, err := model.Execute(scenario, nonMemArguments...)
-					t.R.NoError(err)
-
-					WaitForSGEToTerminate(getGridNameIdentifier(model))
-
-					testingDetails := NonMemTestingDetails{
-						OutputDir: filepath.Join(scenario.Workpath, model.identifier),
-						Model:     model,
-					}
-
-					AssertNonMemCompleted(t, testingDetails)
-					AssertNonMemCreatedOutputFiles(t, testingDetails)
-					AssertNonMemCleanedUpFiles(t, testingDetails)
-					AssertContainsBBIScript(t, testingDetails)
-				})
-			}
-		})
-	}
+	checkGridExecution(tt, "sge", tests, WaitForSGEToTerminate)
 }
 
 func TestBbiCompletesParallelSGEExecution(tt *testing.T) {
@@ -75,62 +28,13 @@ func TestBbiCompletesParallelSGEExecution(tt *testing.T) {
 		tt.Skip("Skipping SG Parallel execution as it's not enabled")
 	}
 
-	tests := []struct {
-		name string
-	}{
-		{name: "acop"},
-		{name: "ctl_test"},
-		{name: "leading-path-with space"},
+	tests := []string{
+		"acop",
+		"ctl_test",
+		"leading-path-with space",
 	}
 
-	for _, test := range tests {
-		tt.Run("SGE "+test.name, func(tt *testing.T) {
-			t := wrapt.WrapT(tt)
-
-			scenario := InitializeScenario(t, test.name)
-
-			// Test shouldn't take longer than 5 min in total
-			// TODO use the context downstream in a runModel function
-			ctx, cancel := context.WithTimeout(context.Background(), 30*time.Minute)
-			defer cancel()
-
-			scenario.Prepare(t, ctx)
-
-			for _, m := range scenario.models {
-				t.Run(m.identifier, func(t *wrapt.T) {
-					nonMemArguments := []string{
-						"-d",
-						"nonmem",
-						"run",
-						"sge",
-						"--nm_version",
-						os.Getenv("NMVERSION"),
-						"--parallel=true",
-						"--mpi_exec_path",
-						os.Getenv("MPIEXEC_PATH"),
-						"--threads",
-						"2",
-					}
-
-					_, err := m.Execute(scenario, nonMemArguments...)
-					t.R.NoError(err)
-
-					WaitForSGEToTerminate(getGridNameIdentifier(m))
-
-					testingDetails := NonMemTestingDetails{
-						OutputDir: filepath.Join(scenario.Workpath, m.identifier),
-						Model:     m,
-					}
-
-					AssertNonMemCompleted(t, testingDetails)
-					AssertNonMemCreatedOutputFiles(t, testingDetails)
-					AssertNonMemCleanedUpFiles(t, testingDetails)
-					AssertContainsBBIScript(t, testingDetails)
-					AssertNonMemOutputContainsParafile(t, testingDetails)
-				})
-			}
-		})
-	}
+	checkParallelGridExecution(tt, "sge", tests, WaitForSGEToTerminate)
 }
 
 func WaitForSGEToTerminate(gridNameIdentifier string) {
@@ -149,12 +53,4 @@ func CountOfPendingJobs(gridNameIdentifier string) int {
 	})
 
 	return len(jobs)
-}
-
-func getGridNameIdentifier(model Model) string {
-	if envValue := os.Getenv("BBI_GRID_NAME_PREFIX"); envValue != "" {
-		return envValue + "_Run_" + model.identifier
-	} else {
-		return "Run_" + model.identifier
-	}
 }
